src/backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
sqlalchemy
psycopg2-binary
python-jose
cryptography
requests
apscheduler